# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/admin/clearinghouse/credentials/{clearinghouse}": {
      "put": {
        "operationId": "ClearinghouseCredentialController_upsert",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Upsert credentials for one clearinghouse. Plaintext is encrypted at rest with AES-256-GCM.",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/clearinghouse/credentials/{id}": {
      "delete": {
        "operationId": "ClearinghouseCredentialController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Remove a credential row.",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/clearinghouse/credentials/{id}/test": {
      "post": {
        "operationId": "ClearinghouseCredentialController_test",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Test credentials by minting an access token. Updates last_verified_at on success/failure.",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/disputes": {
      "post": {
        "operationId": "DisputeController_submit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Submit a \"this rule is wrong\" dispute",
        "tags": [
          "disputes"
        ]
      }
    },
    "/v1/disputes/{id}/resolve-right": {
      "post": {
        "operationId": "DisputeController_resolveRight",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveRightDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Analyst confirms our rule is correct",
        "tags": [
          "disputes"
        ]
      }
    },
    "/v1/disputes/{id}/resolve-wrong": {
      "post": {
        "operationId": "DisputeController_resolveWrong",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveWrongDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Analyst confirms our rule is wrong; spawns extraction_candidate",
        "tags": [
          "disputes"
        ]
      }
    },
...
    "/v1/admin/invites": {
      "post": {
        "operationId": "InviteIssueController_issue",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Issue a magic-link invite for an existing app_user",
        "tags": [
          "admin"
        ]
      },
      "get": {
        "operationId": "InviteIssueController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List recent invites for the calling tenant",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/invites/{id}": {
      "delete": {
        "operationId": "InviteIssueController_revoke",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Revoke an outstanding invite",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/invite/redeem": {
      "post": {
        "operationId": "InviteRedeemController_redeem",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedeemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Redeem a magic-link invite. Anonymous + rate-limited.",
        "tags": [
          "invite"
        ]
      }
    },
...
  },
  "info": {
    "title": "Billing Rules Platform API",
    "description": "Pre-flight a claim against payer × state × code rules.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "org-id": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Org-Id"
      }
    },
    "schemas": {
...
      "UpsertDto": {
        "type": "object",
        "properties": {}
      },
...
      "SubmitDto": {
        "type": "object",
        "properties": {}
      },
      "ResolveRightDto": {
        "type": "object",
        "properties": {}
      },
      "ResolveWrongDto": {
        "type": "object",
        "properties": {}
      },
...
      "IssueDto": {
        "type": "object",
        "properties": {}
      },
      "RedeemDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}